MOB-8494: Add NO_ATTACH flag to interface filter specification
This commit allows interface filters to specify a NO_ATTACH flag to indicate
that ceusbkwrapper should not accept control of certain devices. If the flag
is specified in an interface filter, then ceusbkwrapper will not accept control
of any devices that the filter matches against. This is useful when particular
third-party USB device drivers refuse to operate on a device simultaneously with
ceusbkwrapper.