MOB-8787: Adding ordering to interface filters
This commit adds an optional priority field to the start of the interface
filter string. A priority of 0 represents highest priority, whilst MAXDWORD
represents lowest (no) priority, which is the default if the priority field is
omitted. Interface filters which specify a priority are evaluated in priority
order, and before those with no priority. As the priority field is optional,
previous interface filter strings remain compatible.

This commit also refactors the interface filter string parsing code, and
corrects a minor error in a couple of ERROR_MSG() statements in UserBuffer.cpp.