MOB-8892: Try to find drivers for all interfaces with matching filters before entire device.
This patch updates ceusbkwrapper to search for drivers for each interface
with a matching filter, before attempting to load a driver for the entire
device. Loading a driver for the entire device is considered a final last-ditch
attempt, and so should only be attempted once each interface has been exhausted.